adapters/AutoGluon/requirements.txt
autogluon
grpcio
dill
jinja2
protobuf==3.17.3
pillow
sktime==0.12.1
